Update transaction creation to iota.rs changes and prepare for offline signing
