+++ src/account/operations/transfer/mod.rs
@@ -3,6 +3,7 @@
 
 // transfer or transaction?
 
+mod build_transaction;
 pub(crate) mod high_level;
 mod input_selection;
 mod options;
@@ -14,14 +15,12 @@ use std::time::{SystemTime, UNIX_EPOCH};
 
 use iota_client::{
     bee_message::{
-        input::INPUT_COUNT_RANGE,
-        output::{ByteCostConfig, Output, OUTPUT_COUNT_RANGE},
+        output::{ByteCostConfig, Output},
         payload::transaction::{TransactionId, TransactionPayload},
         MessageId,
     },
     secret::types::InputSigningData,
 };
-use packable::bounded::TryIntoBoundedU16Error;
 use serde::Serialize;
 
 pub use self::options::{RemainderValueStrategy, TransferOptions};
@@ -29,10 +28,9 @@ use crate::account::{
     handle::AccountHandle,
     operations::syncing::SyncOptions,
     types::{InclusionState, Transaction},
-    AddressGenerationOptions,
 };
 #[cfg(feature = "events")]
-use crate::events::types::{AddressData, TransferProgressEvent, WalletEvent};
+use crate::events::types::{TransferProgressEvent, WalletEvent};
 
 /// The result of a transfer, message_id is an option because submitting the transaction could fail
 #[derive(Debug, Serialize)]
@@ -101,122 +99,14 @@ impl AccountHandle {
         byte_cost_config: &ByteCostConfig,
     ) -> crate::Result<TransferResult> {
         log::debug!("[TRANSFER] send");
-        // Check if the outputs have enough amount to cover the storage deposit
-        for output in &outputs {
-            output.verify_storage_deposit(byte_cost_config)?;
-        }
 
-        // validate amounts
-        if !OUTPUT_COUNT_RANGE.contains(&(outputs.len() as u16)) {
-            return Err(crate::Error::BeeMessage(
-                iota_client::bee_message::Error::InvalidOutputCount(TryIntoBoundedU16Error::Truncated(outputs.len())),
-            ));
-        }
+        let prepared_transaction_data = self.prepare_transaction(outputs, options, byte_cost_config).await?;
 
-        let custom_inputs: Option<Vec<InputSigningData>> = {
-            if let Some(options) = options.clone() {
-                // validate inputs amount
-                if let Some(inputs) = &options.custom_inputs {
-                    if !INPUT_COUNT_RANGE.contains(&(inputs.len() as u16)) {
-                        return Err(crate::Error::BeeMessage(
-                            iota_client::bee_message::Error::InvalidInputCount(TryIntoBoundedU16Error::Truncated(
-                                inputs.len(),
-                            )),
-                        ));
-                    }
-                    let account = self.read().await;
-                    let mut input_outputs = Vec::new();
-                    for output_id in inputs {
-                        match account.unspent_outputs().get(output_id) {
-                            Some(output) => input_outputs.push(output.input_signing_data()?),
-                            None => {
-                                return Err(crate::Error::CustomInputError(format!(
-                                    "Custom input {} not found in unspent outputs",
-                                    output_id
-                                )));
-                            }
-                        }
-                    }
-                    Some(input_outputs)
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        };
-
-        let remainder_address = match &options {
-            Some(options) => {
-                match &options.remainder_value_strategy {
-                    RemainderValueStrategy::ReuseAddress => {
-                        // select_inputs will select an address from the inputs if it's none
-                        None
-                    }
-                    RemainderValueStrategy::ChangeAddress => {
-                        let remainder_address = self
-                            .generate_addresses(
-                                1,
-                                Some(AddressGenerationOptions {
-                                    internal: true,
-                                    ..Default::default()
-                                }),
-                            )
-                            .await?
-                            .first()
-                            .expect("Didn't generate an address")
-                            .clone();
-                        #[cfg(feature = "events")]
-                        {
-                            let account_index = self.read().await.index;
-                            self.event_emitter.lock().await.emit(
-                                account_index,
-                                WalletEvent::TransferProgress(
-                                    TransferProgressEvent::GeneratingRemainderDepositAddress(AddressData {
-                                        address: remainder_address.address.to_bech32(),
-                                    }),
-                                ),
-                            );
-                        }
-                        Some(remainder_address.address().inner)
-                    }
-                    RemainderValueStrategy::CustomAddress(address) => Some(address.address().inner),
-                }
-            }
-            None => None,
-        };
-
-        let selected_transaction_data = self
-            .select_inputs(outputs, custom_inputs, remainder_address, byte_cost_config)
-            .await?;
-        // can we unlock the outputs in a better way if the transaction creation fails?
-        let prepared_transaction_data = match self
-            .prepare_transaction(
-                selected_transaction_data.inputs.clone(),
-                selected_transaction_data.outputs.clone(),
-                options,
-            )
-            .await
-        {
+        let transaction_payload = match self.sign_tx_essence(&prepared_transaction_data).await {
             Ok(res) => res,
             Err(err) => {
                 // unlock outputs so they are available for a new transaction
-                self.unlock_inputs(selected_transaction_data.inputs).await?;
-                return Err(err);
-            }
-        };
-        let transaction_payload = match self
-            .sign_tx_essence(
-                prepared_transaction_data.essence,
-                prepared_transaction_data.input_signing_data_entries,
-                selected_transaction_data.remainder_output,
-            )
-            .await
-        {
-            Ok(res) => res,
-            Err(err) => {
-                // unlock outputs so they are available for a new transaction
-                self.unlock_inputs(selected_transaction_data.inputs).await?;
+                self.unlock_inputs(prepared_transaction_data.inputs_data).await?;
                 return Err(err);
             }
         };
