+++ src/account/types/address.rs
@@ -75,7 +75,8 @@ impl AddressWrapper {
         self.inner.to_bech32(&self.bech32_hrp)
     }
 
-    pub(crate) fn bech32_hrp(&self) -> &str {
+    /// Get the bech32 human readable part
+    pub fn bech32_hrp(&self) -> &str {
         &self.bech32_hrp
     }
 }
@@ -84,6 +85,6 @@ pub fn parse_bech32_address<A: AsRef<str>>(address: A) -> crate::Result<AddressW
     let address = address.as_ref();
     let mut tokens = address.split('1');
     let hrp = tokens.next().ok_or(crate::Error::InvalidAddress)?;
-    let (_bech32_hrp, address) = iota_client::bee_message::address::Address::try_from_bech32(address)?;
+    let (_bech32_hrp, address) = Address::try_from_bech32(address)?;
     Ok(AddressWrapper::new(address, hrp.to_string()))
 }
